Fixed text editor layout
